+++ docker/mongodb-enterprise-tests/kubetester/kubetester.py
@@ -55,12 +55,33 @@ def is_multi_cluster():
     return len(os.getenv("MEMBER_CLUSTERS", "")) > 0
 
 
+# TODO: Rename to is_default_architecture_static
 def is_static_containers_architecture() -> bool:
     return os.getenv("MDB_DEFAULT_ARCHITECTURE", "non-static") == "static"
 
 
+# TODO: Rename to get_default_architecture
 def get_static_containers_architecture() -> str:
     return "static" if is_static_containers_architecture() else "non-static"
+
+
+def assert_statefulset_architecture(statefulset: client.V1StatefulSet, architecture: str):
+    """
+    Asserts that the statefulset is configured with the expected architecture.
+    """
+    agent_container = next((c for c in statefulset.spec.template.spec.containers if c.name == "mongodb-agent"), None)
+    if architecture == "non-static":
+        # In non-static architecture expect agent container to not be present
+        assert agent_container is None
+    else:
+        # In static architecture we expect agent container to be present
+        # and contain static environment variable which
+        # instructs the agent launcher script to not download binaries
+        assert agent_container is not None
+        static_env_var = next(
+            (env for env in agent_container.env if env.name == "MDB_STATIC_CONTAINERS_ARCHITECTURE"), None
+        )
+        assert static_env_var.value == "true"
 
 
 skip_if_static_containers = pytest.mark.skipif(